chore(tags): remove dead filterChatsByTags, relocate UNTAGGED
Tag filtering runs server-side now, leaving the client-side pure
filter `filterChatsByTags` orphaned — only its own unit test still
referenced it. The `UNTAGGED` sentinel exported from the same module
is still live (the navigation Tags section and the tag-selection
toggler import it).

Move `UNTAGGED` to its own small module (web/src/tags/untagged.ts)
with unchanged value and semantics, repoint the live importers, then
delete the dead function and its test.

Closes #168
