+++ web/src/tags/untagged.ts
@@ -1,0 +1,6 @@
+/**
+ * The sentinel selection entry for the `Untagged` group (Chats holding zero
+ * Tags). Matches the API/reader convention where an empty-string entry selects
+ * the no-Tag bucket, so the navigation panel and the wire form read the same.
+ */
+export const UNTAGGED = "";
